app/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
mlflow==2.20.0
